Refactor: Extract deadline and logging utilities to src/analysis/utils/deadline.rs (#268)
Extract ~820 lines of deadline handling, logging, and randomisation
utilities from implementation.rs to a dedicated module.

Functions moved:
- Deadline utilities: calculate_effective_timeout_ms, build_deadline,
  deadline_passed, calculate_gpu_batch_timeout
- Logging: log_analysis_start, log_analysis_timeout
- Randomisation: derive_seed, shuffle_slice, shuffle_within_top_k
- Environment helpers: parse_input_index, source_input_index_bias_from_env,
  focus_unused_observations_from_env, order_eligible_sources
- Types: OrderedNeuron, deadline constants
- Test infrastructure: deadline_override module

This is part of the larger effort to refactor the implementation.rs
monolith (Issue #185).
