+++ src/analysis/implementation.rs
@@ -29,18 +29,25 @@ use crate::analysis::utils::{
     suppress_mesa_warnings_if_requested, verbose_enabled, MemoryTier, DEFAULT_GPU_BATCH_SIZE,
     HIGH_PERF_GPU_BATCH_SIZE, LOW_MEMORY_GPU_BATCH_SIZE,
 };
+
+// Import deadline handling and logging utilities from dedicated module (Issue #268)
+#[cfg(test)]
+use crate::analysis::utils::deadline_override;
+use crate::analysis::utils::{
+    build_deadline, calculate_gpu_batch_timeout, deadline_passed, log_analysis_start,
+    log_analysis_timeout, order_eligible_sources, parse_input_index, shuffle_slice,
+    shuffle_within_top_k, OrderedNeuron, GPU_QUEUE_TIMEOUT_MAX_SECS,
+};
 use bytemuck::{Pod, Zeroable};
 use crossbeam_channel::{bounded, Receiver, Sender};
 use once_cell::sync::OnceCell;
-use rand::seq::SliceRandom;
-use rand::{rngs::StdRng, Rng, SeedableRng};
 use rayon::prelude::*;
 use std::cmp::Ordering;
 use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::sync::{mpsc, Arc, Mutex};
 use std::thread::{self, JoinHandle};
-use std::time::{Duration, SystemTime};
+use std::time::Duration;
 use wgpu::util::DeviceExt;
 
 const EPSILON: f32 = 1e-8;
@@ -249,15 +256,7 @@ fn constant_source_effect_threshold_from_env() -> Option<f32> {
     }
 }
 
-/// Minimum timeout (seconds) for the GPU work queue waiting for a response from the GPU thread.
-/// This is the OUTER timeout - if the GPU thread doesn't respond within this time,
-/// the queue gives up and returns an error.
-/// For large datasets (>1GB Parquet files), batch evaluation may take longer than 60 seconds.
-const GPU_QUEUE_TIMEOUT_MIN_SECS: u64 = 60;
-
-/// Maximum timeout (seconds) for GPU batch operations.
-/// Even with large datasets, if the GPU hasn't responded in 5 minutes, something is wrong.
-const GPU_QUEUE_TIMEOUT_MAX_SECS: u64 = 300;
+// GPU_QUEUE_TIMEOUT_MIN_SECS and GPU_QUEUE_TIMEOUT_MAX_SECS moved to utils/deadline.rs (Issue #268)
 
 /// Buffer map timeout is always 5 seconds shorter than queue timeout to avoid race conditions.
 /// If both timeouts are the same, the queue might timeout before the GPU thread
@@ -290,39 +289,7 @@ const GPU_MAX_BATCH_ALLOC_BYTES: usize = 256 * 1024 * 1024; // 256MB
 /// GPU device creation should be fast; if it takes longer, something is wrong.
 const GPU_INIT_TIMEOUT_SECS: u64 = 30;
 
-/// Calculate adaptive GPU batch timeout based on remaining deadline.
-///
-/// For large datasets (>1GB Parquet files), GPU batch evaluations may take
-/// longer than the minimum 60 seconds. This function calculates a reasonable
-/// timeout based on:
-/// - Minimum: GPU_QUEUE_TIMEOUT_MIN_SECS (60s) - catches unresponsive GPU
-/// - Maximum: GPU_QUEUE_TIMEOUT_MAX_SECS (5 min) - prevents infinite waits
-/// - If deadline is available: uses up to half remaining time (capped at max)
-fn calculate_gpu_batch_timeout(deadline: &Option<std::time::SystemTime>) -> Duration {
-    let min_timeout = Duration::from_secs(GPU_QUEUE_TIMEOUT_MIN_SECS);
-    let max_timeout = Duration::from_secs(GPU_QUEUE_TIMEOUT_MAX_SECS);
-
-    match deadline {
-        Some(dl) => {
-            if let Ok(remaining) = dl.duration_since(std::time::SystemTime::now()) {
-                // Use half of remaining time, but capped between min and max
-                let half_remaining = remaining / 2;
-                if half_remaining < min_timeout {
-                    min_timeout
-                } else if half_remaining > max_timeout {
-                    max_timeout
-                } else {
-                    half_remaining
-                }
-            } else {
-                // Deadline already passed - use minimum
-                min_timeout
-            }
-        }
-        // No deadline - use maximum
-        None => max_timeout,
-    }
-}
+// calculate_gpu_batch_timeout moved to utils/deadline.rs (Issue #268)
 
 /// Poll the GPU device until it has no work in flight, or a timeout is reached.
 ///
@@ -646,518 +613,8 @@ fn log_gpu_info_once(
     });
 }
 
-/// Constants for deadline calculation - shared between `calculate_effective_timeout_ms` and `build_deadline`.
-const DEFAULT_DURATION_MS: u64 = 600_000; // 10 minutes (10 * 60 * 1000)
-const YEAR_2000_MS: u64 = 946_684_800_000;
-const MIN_DURATION_MS: u64 = 3_000; // 3 seconds
-const MAX_DURATION_MS: u64 = 3_600_000; // 1 hour (60 * 60 * 1000)
-
-/// Calculate the effective timeout duration in milliseconds.
-///
-/// This function applies the same logic as `build_deadline`:
-/// 1. Converts absolute timestamps (values >= year 2000 in ms) to relative durations
-/// 2. Clamps values outside the 3-second to 1-hour range to the 10-minute default
-///
-/// Returns `None` if the deadline is in the past (for absolute timestamps), or
-/// `Some(effective_duration_ms)` otherwise.
-///
-/// Used by both `build_deadline` (to create the SystemTime) and `log_analysis_start`
-/// (to display the effective timeout to users).
-fn calculate_effective_timeout_ms(deadline_ms: Option<u64>) -> Option<u64> {
-    let target_ms = deadline_ms.unwrap_or(DEFAULT_DURATION_MS);
-
-    // Heuristic: if the value is less than year 2000 in milliseconds,
-    // treat it as a relative duration. Otherwise, it's likely an absolute timestamp
-    // from the calling code, so convert it to a relative duration.
-    let relative_ms = if target_ms < YEAR_2000_MS {
-        // Small value - treat as relative duration (milliseconds from now)
-        target_ms
-    } else {
-        // Large value - likely an absolute timestamp from calling code.
-        // Convert to relative duration by subtracting current time.
-        let now_ms = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .ok()?
-            .as_millis() as u64;
-
-        // If the timestamp is in the past, return None (deadline already passed)
-        if target_ms <= now_ms {
-            return None;
-        }
-
-        // Calculate relative duration
-        target_ms - now_ms
-    };
-
-    // Validate duration bounds: minimum 3 seconds, maximum 1 hour
-    // If invalid, default to 10 minutes (expected typical value)
-    // NOTE: If these warnings appear, it's a bug in the calling code (NEAT-AI or GRQ)
-    // that should be fixed to pass valid timeout values.
-    let validated_ms = if relative_ms < MIN_DURATION_MS {
-        eprintln!(
-            "⚠️  [NEAT-AI-Discovery] BUG: analysis_deadline_ms ({:.1}s) is less than minimum (3s). \
-             Using default 10 minute timeout. Please fix the calling code (NEAT-AI/GRQ) to pass a valid timeout.",
-            relative_ms as f64 / 1000.0
-        );
-        DEFAULT_DURATION_MS
-    } else if relative_ms > MAX_DURATION_MS {
-        eprintln!(
-            "⚠️  [NEAT-AI-Discovery] BUG: analysis_deadline_ms ({:.1}s) exceeds maximum (1 hour). \
-             Using default 10 minute timeout. Please fix the calling code (NEAT-AI/GRQ) to pass a valid timeout.",
-            relative_ms as f64 / 1000.0
-        );
-        DEFAULT_DURATION_MS
-    } else {
-        relative_ms
-    };
-
-    Some(validated_ms)
-}
-
-fn parse_input_index(uuid: &str) -> Option<usize> {
-    uuid.strip_prefix("input-")?.parse::<usize>().ok()
-}
-
-/// Derive a stable per-context seed from a base seed.
-///
-/// We deliberately avoid `Hash` here because Rust's hasher is intentionally
-/// randomised between processes.
-fn derive_seed(base_seed: u64, context: &str, salt: u64) -> u64 {
-    // FNV-1a 64-bit
-    let mut hash: u64 = 0xcbf29ce484222325 ^ base_seed ^ salt;
-    for b in context.as_bytes() {
-        hash ^= *b as u64;
-        hash = hash.wrapping_mul(0x100000001b3);
-    }
-    hash
-}
-
-/// Shuffle a slice, optionally deterministically.
-///
-/// When `seed` is `None`, uses non-deterministic randomness (best for production
-/// timeout runs where we want coverage to drift over time).
-fn shuffle_slice<T>(items: &mut [T], seed: Option<u64>, context: &str) {
-    if items.len() <= 1 {
-        return;
-    }
-
-    match seed {
-        Some(base) => {
-            let derived = derive_seed(base, context, 0x9e3779b97f4a7c15);
-            let mut rng = StdRng::seed_from_u64(derived);
-            items.shuffle(&mut rng);
-        }
-        None => {
-            let mut rng = rand::thread_rng();
-            items.shuffle(&mut rng);
-        }
-    }
-}
-
-/// Diversify a best-first candidate list by shuffling within the top-K prefix.
-///
-/// Rationale (Jan 2026):
-/// - In production, discovery runs are deadline-constrained and repeated over time.
-/// - The controller (TypeScript) caches failed candidates and will not re-attempt them
-///   until the training data changes.
-/// - If Rust always returns the same top few candidates, categories can starve because
-///   those candidates become permanently skipped for the day.
-///
-/// Shuffling within the top-K keeps us focused on high-quality candidates while still
-/// drifting over time, improving long-run coverage without returning low-quality tails.
-fn shuffle_within_top_k<T>(items: &mut [T], seed: Option<u64>, context: &str, top_k: usize) {
-    if items.len() <= 1 || top_k <= 1 {
-        return;
-    }
-    let k = top_k.min(items.len());
-    shuffle_slice(&mut items[..k], seed, context);
-}
-
-#[cfg(test)]
-mod tests_shuffle_within_top_k {
-    use super::shuffle_within_top_k;
-
-    #[test]
-    fn deterministic_with_seed() {
-        let mut a = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
-        let mut b = a.clone();
-        shuffle_within_top_k(&mut a[..], Some(123), "ctx", 5);
-        shuffle_within_top_k(&mut b[..], Some(123), "ctx", 5);
-        assert_eq!(a, b);
-    }
-
-    #[test]
-    fn only_shuffles_prefix() {
-        let mut items = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
-        shuffle_within_top_k(&mut items[..], Some(123), "ctx", 3);
-        // Suffix must remain untouched.
-        assert_eq!(items[3..], [4, 5, 6, 7, 8]);
-    }
-
-    #[test]
-    fn preserves_multiset() {
-        let mut items = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
-        let mut before = items.clone();
-        before.sort_unstable();
-        shuffle_within_top_k(&mut items[..], Some(999), "ctx", 6);
-        items.sort_unstable();
-        assert_eq!(items, before);
-    }
-
-    #[test]
-    fn no_op_when_top_k_is_zero() {
-        let mut items = [1u32, 2, 3, 4].to_vec();
-        let before = items.clone();
-        shuffle_within_top_k(&mut items[..], Some(1), "ctx", 0);
-        assert_eq!(items, before);
-    }
-}
-
-/// Optional input-index bias for source ordering.
-///
-/// This is primarily a production knob for timeout-constrained runs. If you are
-/// frequently appending new inputs (e.g., feature engineering), you may want to
-/// bias discovery toward higher input indices so newer inputs are evaluated
-/// earlier, while still allowing older inputs to be discovered over time.
-///
-/// Controlled via `NEAT_AI_DISCOVERY_SOURCE_INPUT_INDEX_BIAS`:
-/// - unset / empty: disabled (pure shuffle)
-/// - `> 0`: enabled (weighted random permutation; higher = stronger bias to the end)
-fn source_input_index_bias_from_env() -> Option<f64> {
-    let raw = std::env::var("NEAT_AI_DISCOVERY_SOURCE_INPUT_INDEX_BIAS").ok()?;
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    match trimmed.parse::<f64>() {
-        Ok(v) if v.is_finite() && v > 0.0 => Some(v),
-        _ => {
-            if verbose_enabled() {
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] Ignoring invalid NEAT_AI_DISCOVERY_SOURCE_INPUT_INDEX_BIAS={trimmed:?} (expected a finite number > 0)"
-                );
-            }
-            None
-        }
-    }
-}
-
-/// Check if discovery should focus on unused observations (Issue #182).
-///
-/// When enabled via `NEAT_AI_DISCOVERY_FOCUS_UNUSED_OBSERVATIONS=1`, input neurons
-/// that have NO existing outgoing synapses are prioritised in the source ordering.
-/// This is useful when new observations have been added to the training data and
-/// the user wants discovery to focus on connecting these new inputs first.
-///
-/// Values that enable the feature: "1", "true", "yes" (case-insensitive)
-/// Values that disable the feature: unset, empty, "0", "false", "no"
-pub fn focus_unused_observations_from_env() -> bool {
-    let raw = match std::env::var("NEAT_AI_DISCOVERY_FOCUS_UNUSED_OBSERVATIONS") {
-        Ok(v) => v,
-        Err(_) => return false,
-    };
-    let trimmed = raw.trim().to_lowercase();
-    matches!(trimmed.as_str(), "1" | "true" | "yes")
-}
-
-/// Order eligible sources for evaluation.
-///
-/// - Always respects forward-only candidate constraints (caller must filter by index).
-/// - Default behaviour is a pure random shuffle (no special casing for input indices).
-/// - If `NEAT_AI_DISCOVERY_SOURCE_INPUT_INDEX_BIAS` is set, input sources are ordered
-///   by a weighted random permutation favouring higher input indices.
-/// - If `NEAT_AI_DISCOVERY_FOCUS_UNUSED_OBSERVATIONS=1` is set, input neurons with NO
-///   existing outgoing synapses are prioritised (moved to the front) before any other
-///   ordering is applied (Issue #182).
-fn order_eligible_sources(
-    eligible_sources: &mut Vec<&OrderedNeuron>,
-    seed: Option<u64>,
-    context: &str,
-    creature_input_count: usize,
-    used_inputs: Option<&HashSet<String>>,
-) {
-    if eligible_sources.len() <= 1 {
-        return;
-    }
-
-    // Issue #182: When NEAT_AI_DISCOVERY_FOCUS_UNUSED_OBSERVATIONS=1 is set,
-    // prioritise input neurons that have NO existing outgoing synapses.
-    // These "unused observations" are moved to the front of the list.
-    if focus_unused_observations_from_env() {
-        if let Some(used) = used_inputs {
-            // Partition: unused inputs first, then others
-            // An input is "unused" if it's an input neuron AND not in the used_inputs set
-            let (mut unused_inputs, mut others): (Vec<_>, Vec<_>) = eligible_sources
-                .drain(..)
-                .partition(|n| parse_input_index(&n.uuid).is_some() && !used.contains(&n.uuid));
-
-            // Log when focusing on unused observations
-            if verbose_enabled() && !unused_inputs.is_empty() {
-                let unused_count = unused_inputs.len();
-                let used_count = others
-                    .iter()
-                    .filter(|n| parse_input_index(&n.uuid).is_some())
-                    .count();
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] Focus unused observations: prioritising {} unused inputs over {} used inputs and {} other sources",
-                    unused_count,
-                    used_count,
-                    others.len() - used_count
-                );
-            }
-
-            // Shuffle each partition separately, then concatenate
-            shuffle_slice(&mut unused_inputs, seed, &format!("{context}:unused"));
-            shuffle_slice(&mut others, seed, &format!("{context}:others"));
-
-            eligible_sources.extend(unused_inputs);
-            eligible_sources.extend(others);
-            return;
-        }
-    }
-
-    let Some(bias) = source_input_index_bias_from_env() else {
-        shuffle_slice(eligible_sources, seed, context);
-        return;
-    };
-
-    // Weighted permutation via exponential race:
-    // t = -ln(U) / w, smaller t comes first.
-    let max_input_index = creature_input_count.saturating_sub(1) as f64;
-    let denom = (max_input_index + 1.0).max(1.0);
-
-    let mut rng = match seed {
-        Some(base) => {
-            let derived = derive_seed(base, context, 0x243f6a8885a308d3);
-            StdRng::seed_from_u64(derived)
-        }
-        None => {
-            // Use a random seed then a deterministic RNG instance so we can reuse the same
-            // code path without fighting trait object ergonomics.
-            let seed: u64 = rand::thread_rng().gen();
-            StdRng::seed_from_u64(seed)
-        }
-    };
-
-    let mut keyed: Vec<(f64, &OrderedNeuron)> = Vec::with_capacity(eligible_sources.len());
-    for &n in eligible_sources.iter() {
-        let weight = if let Some(i) = parse_input_index(&n.uuid) {
-            // Normalise to (0, 1] based on input index, then apply power bias.
-            // Epsilon keeps weight > 0 even for i=0 with high bias.
-            let x = ((i as f64) + 1.0) / denom;
-            (1e-6 + x).powf(bias)
-        } else {
-            1.0
-        };
-
-        let u: f64 = rng.gen::<f64>().max(f64::MIN_POSITIVE);
-        let t = -u.ln() / weight.max(1e-12);
-        keyed.push((t, n));
-    }
-
-    keyed.sort_by(|a, b| a.0.partial_cmp(&b.0).unwrap_or(Ordering::Equal));
-    eligible_sources.clear();
-    eligible_sources.extend(keyed.into_iter().map(|(_, n)| n));
-}
-
-fn build_deadline(deadline_ms: Option<u64>) -> Option<SystemTime> {
-    // Treat deadline_ms as a relative duration (milliseconds from now), not an absolute timestamp.
-    // The calling code (TypeScript) calculates this as Date.now() + duration, but we want to treat
-    // it as a duration to avoid issues with clock skew and to match the expected semantics.
-    // If the calling code passes an absolute timestamp, we need to convert it to a relative duration.
-    // If None is passed, apply default 10 minute timeout to prevent runaway analysis.
-    calculate_effective_timeout_ms(deadline_ms)
-        .and_then(|validated_ms| SystemTime::now().checked_add(Duration::from_millis(validated_ms)))
-}
-
-fn deadline_passed(deadline: &Option<SystemTime>) -> bool {
-    #[cfg(test)]
-    {
-        if let Some(value) = deadline_override::next_override_value() {
-            return value;
-        }
-    }
-
-    matches!(deadline, Some(limit) if SystemTime::now() >= *limit)
-}
-
-/// Log analysis start information including deadline and focus neuron count.
-/// This provides visibility into timeout configuration without requiring verbose mode.
-fn log_analysis_start(
-    analysis_type: &str,
-    deadline_ms: Option<u64>,
-    focus_count: usize,
-    shuffled_order: &[String],
-) {
-    // Calculate the effective deadline duration using the same logic as build_deadline.
-    // This ensures the logged timeout matches what's actually used.
-    let deadline_duration_ms =
-        calculate_effective_timeout_ms(deadline_ms).unwrap_or(DEFAULT_DURATION_MS);
-    let deadline_secs = deadline_duration_ms as f64 / 1000.0;
-
-    // Debug: Log the raw deadline_ms value if verbose to help diagnose timeout issues
-    if verbose_enabled() {
-        if let Some(raw_ms) = deadline_ms {
-            let now_ms = SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)
-                .map(|d| d.as_millis() as u64)
-                .unwrap_or(0);
-            if raw_ms >= YEAR_2000_MS {
-                // Absolute timestamp
-                let elapsed_secs =
-                    now_ms.saturating_sub(raw_ms.saturating_sub(deadline_duration_ms)) as f64
-                        / 1000.0;
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] {analysis_type} deadline_ms={raw_ms} (absolute timestamp), \
-                     {elapsed_secs:.1}s elapsed since timeout was set, {deadline_secs:.1}s remaining"
-                );
-            } else {
-                // Relative duration
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] {analysis_type} deadline_ms={raw_ms} (relative duration)"
-                );
-            }
-        }
-    }
-
-    // Warn if timeout is very short - likely means focus selection took most of the allotted time
-    const MIN_USEFUL_TIMEOUT_SECS: f64 = 60.0; // 1 minute minimum for useful analysis
-    if deadline_secs < MIN_USEFUL_TIMEOUT_SECS {
-        eprintln!(
-            "💡  [NEAT-AI-Discovery]: Only {deadline_secs:.1}s remaining for {analysis_type} analysis. \
-             Focus selection may have consumed most of the timeout. \
-             Consider increasing discoveryAnalysisTimeoutMinutes."
-        );
-    }
-
-    // Format the timeout nicely
-    let timeout_str = if deadline_secs >= 60.0 {
-        let minutes = deadline_secs / 60.0;
-        format!("{minutes:.1} minutes")
-    } else {
-        format!("{deadline_secs:.1} seconds")
-    };
-
-    eprintln!(
-        "[NEAT-AI-Discovery] Starting {analysis_type} analysis: {focus_count} focus neurons, timeout: {timeout_str}"
-    );
-
-    // Log the shuffled order if verbose mode is enabled
-    if verbose_enabled() && !shuffled_order.is_empty() {
-        let preview: Vec<&str> = shuffled_order.iter().take(5).map(|s| s.as_str()).collect();
-        let extra = shuffled_order.len().saturating_sub(5);
-        let suffix = if extra > 0 {
-            format!("... (+{extra} more)")
-        } else {
-            String::new()
-        };
-        eprintln!("[NEAT-AI-Discovery][verbose] Randomised focus order: {preview:?}{suffix}");
-    }
-}
-
-/// Log when analysis timeout is reached. Always prints (not verbose-only).
-fn log_analysis_timeout(analysis_type: &str, completed_count: usize, total_count: usize) {
-    eprintln!(
-        "[NEAT-AI-Discovery] {analysis_type} analysis reached timeout. Completed {completed_count}/{total_count} focus neurons. \
-         Returning partial results."
-    );
-}
-
-#[cfg(test)]
-mod deadline_override {
-    use std::cell::RefCell;
-    use std::collections::VecDeque;
-    use std::sync::{Arc, Mutex};
-
-    /// Deadline override state shared across a thread pool.
-    ///
-    /// We store this behind an `Arc` so a single override sequence can be consumed
-    /// by multiple worker threads within a *private* Rayon pool.
-    #[derive(Clone)]
-    struct OverrideState {
-        queue: Arc<Mutex<VecDeque<bool>>>,
-    }
-
-    thread_local! {
-        /// Thread-local handle to the active override state.
-        ///
-        /// IMPORTANT (2 Jan 2026):
-        /// Tests run in parallel and analysis uses Rayon. A global override is unsafe because
-        /// *other tests* running concurrently can observe and consume override values.
-        ///
-        /// By using thread-local state, and explicitly broadcasting it into a test's private
-        /// Rayon pool, we:
-        /// - Avoid deadlocks (no cross-thread lock held for the duration of a test)
-        /// - Avoid cross-test interference under default parallel test execution
-        static OVERRIDE_STATE: RefCell<Option<OverrideState>> = const { RefCell::new(None) };
-    }
-
-    fn set_override_state(state: Option<OverrideState>) {
-        OVERRIDE_STATE.with(|cell| {
-            *cell.borrow_mut() = state;
-        });
-    }
-
-    pub(super) struct DeadlineOverrideGuard {
-        /// When provided, we clear the override from all worker threads on drop.
-        pool: Option<Arc<rayon::ThreadPool>>,
-    }
-
-    impl DeadlineOverrideGuard {
-        /// Install an override sequence for the current thread only.
-        pub(super) fn with_sequence(sequence: Vec<bool>) -> Self {
-            Self::with_sequence_for_pool(sequence, None)
-        }
-
-        /// Install an override sequence for a specific Rayon pool.
-        ///
-        /// This is the safe way to use deadline overrides in multi-threaded tests: create a private
-        /// pool, broadcast the override into that pool, and run analysis inside `pool.install(...)`.
-        pub(super) fn with_sequence_for_pool(
-            sequence: Vec<bool>,
-            pool: Option<Arc<rayon::ThreadPool>>,
-        ) -> Self {
-            let state = OverrideState {
-                queue: Arc::new(Mutex::new(sequence.into_iter().collect::<VecDeque<_>>())),
-            };
-
-            // Always install for the current thread (covers non-parallel code paths).
-            set_override_state(Some(state.clone()));
-
-            // If a pool is provided, broadcast into each worker thread.
-            if let Some(ref p) = pool {
-                p.broadcast(|_| set_override_state(Some(state.clone())));
-            }
-
-            Self { pool }
-        }
-    }
-
-    impl Drop for DeadlineOverrideGuard {
-        fn drop(&mut self) {
-            // Clear on current thread.
-            set_override_state(None);
-            // Clear on pool worker threads (if any).
-            if let Some(ref p) = self.pool {
-                p.broadcast(|_| set_override_state(None));
-            }
-        }
-    }
-
-    pub(super) fn next_override_value() -> Option<bool> {
-        OVERRIDE_STATE.with(|cell| {
-            let state = cell.borrow().clone()?;
-            let mut queue = state
-                .queue
-                .lock()
-                .expect("Deadline override queue should not be poisoned");
-            queue.pop_front()
-        })
-    }
-}
-
-// verbose_enabled, suppress_mesa_warnings_if_requested, and ensure_xdg_runtime_dir
-// moved to crate::analysis::utils (Issue #267)
+// Deadline handling, logging, and randomisation utilities moved to utils/deadline.rs (Issue #268)
+// All functions now imported from crate::analysis::utils
 
 /// Safely create a wgpu Instance, avoiding panics from backend probing.
 ///
@@ -1242,11 +699,7 @@ fn create_wgpu_instance_safely() -> Option<wgpu::Instance> {
 }
 
 // Types moved to shared.rs - using imports from there
-
-struct OrderedNeuron {
-    uuid: String,
-    index: usize,
-}
+// OrderedNeuron moved to utils/deadline.rs (Issue #268)
 
 type RecordCacheLoader = dyn Fn(&str, &str) -> Result<Vec<DiscoverRecord>> + Send + Sync + 'static;
 type CachedNeuronRecords = OnceCell<Arc<Vec<DiscoverRecord>>>;
@@ -9811,7 +9264,7 @@ mod tests_synapses {
     use std::sync::atomic::{AtomicUsize, Ordering as AtomicOrdering};
     use std::sync::{Arc, Barrier};
     use std::thread;
-    use std::time::{Duration, SystemTime};
+    use std::time::Duration;
     use tempfile::tempdir;
 
     /// Helper macro to skip tests that require GPU when no GPU is available.
@@ -9825,266 +9278,7 @@ mod tests_synapses {
         };
     }
 
-    #[test]
-    fn deadline_passed_detects_elapsed_wall_clock_deadline() {
-        // Note: This test does NOT require GPU - it only tests the deadline_passed
-        // function which performs simple time comparisons. Do not add skip_if_no_gpu!()
-        // Use the deadline override mechanism in tests so behaviour is deterministic
-        let _guard = deadline_override::DeadlineOverrideGuard::with_sequence(vec![true, false]);
-
-        let dummy_deadline = Some(SystemTime::now());
-        assert!(
-            deadline_passed(&dummy_deadline),
-            "past deadlines should be treated as expired immediately"
-        );
-
-        assert!(
-            !deadline_passed(&dummy_deadline),
-            "future deadlines should not be marked as expired"
-        );
-
-        assert!(
-            !deadline_passed(&None),
-            "missing deadlines should behave as if no timeout was requested"
-        );
-    }
-
-    #[test]
-    fn build_deadline_handles_absolute_timestamps_and_relative_durations() {
-        // Verify that build_deadline correctly handles both absolute timestamps
-        // (milliseconds since UNIX_EPOCH) and relative durations (milliseconds from now)
-        let now = SystemTime::now();
-        let now_ms = now
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .expect("SystemTime should be after UNIX_EPOCH")
-            .as_millis() as u64;
-
-        // Test 1: Absolute timestamp (large value, >= year 2000)
-        // Create a deadline 10 minutes in the future using absolute timestamp
-        let ten_minutes_ms = 10 * 60 * 1000; // 10 minutes in milliseconds
-        let future_deadline_ms = now_ms + ten_minutes_ms;
-        let deadline = build_deadline(Some(future_deadline_ms));
-
-        assert!(
-            deadline.is_some(),
-            "deadline should be Some when deadline_ms is provided"
-        );
-
-        let deadline_time = deadline.unwrap();
-
-        // The deadline should be approximately 10 minutes in the future
-        // Allow for some small timing variance (up to 1 second)
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "absolute timestamp deadline should be approximately 10 minutes in the future, got {duration:?}"
-            );
-        } else {
-            panic!("deadline should be in the future");
-        }
-
-        // Test 2: Relative duration (small value, < year 2000)
-        // Pass 10 minutes as a relative duration
-        let relative_deadline_ms = ten_minutes_ms; // 10 minutes as relative duration
-        let relative_deadline = build_deadline(Some(relative_deadline_ms));
-        assert!(relative_deadline.is_some());
-        let relative_time = relative_deadline.unwrap();
-        // This should also be approximately 10 minutes in the future
-        if let Ok(duration) = relative_time.duration_since(now) {
-            let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "relative duration deadline should be approximately 10 minutes in the future, got {duration:?}"
-            );
-        } else {
-            panic!("relative deadline should be in the future");
-        }
-
-        // Test 3: Verify that an absolute timestamp in the past returns None
-        // (deadline already passed - no point in creating a deadline)
-        let past_timestamp_ms = 1_700_000_000_000u64; // Jan 2024 (in the past)
-        let past_deadline = build_deadline(Some(past_timestamp_ms));
-        assert!(
-            past_deadline.is_none(),
-            "Past timestamp should return None (deadline already passed)"
-        );
-
-        // Test 4: Verify that a future absolute timestamp is correctly converted to relative duration
-        let future_timestamp_ms = now_ms + ten_minutes_ms; // 10 minutes in the future as absolute timestamp
-        let future_deadline = build_deadline(Some(future_timestamp_ms));
-        assert!(future_deadline.is_some());
-        let future_time = future_deadline.unwrap();
-        // Should be approximately 10 minutes in the future
-        if let Ok(duration) = future_time.duration_since(now) {
-            let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Future absolute timestamp should be converted to relative duration correctly, got {duration:?}"
-            );
-        } else {
-            panic!("Future deadline should be in the future");
-        }
-    }
-
-    #[test]
-    fn build_deadline_validates_duration_bounds() {
-        // Test that build_deadline validates and defaults to 10 minutes for invalid values
-        let now = SystemTime::now();
-        const DEFAULT_DURATION_MS: u64 = 600_000; // 10 minutes
-
-        // Test 1: Duration below minimum (3 seconds) should default to 10 minutes
-        let too_short_ms = 1_000u64; // 1 second
-        let deadline = build_deadline(Some(too_short_ms));
-        assert!(deadline.is_some());
-        let deadline_time = deadline.unwrap();
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            // Should default to 10 minutes (600000 ms)
-            let expected_min = Duration::from_millis(DEFAULT_DURATION_MS) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(DEFAULT_DURATION_MS) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Duration below 3 seconds should default to 10 minutes, got {duration:?}"
-            );
-        } else {
-            panic!("Default deadline should be in the future");
-        }
-
-        // Test 2: Duration above maximum (1 hour) should default to 10 minutes
-        let too_long_ms = 4_000_000u64; // ~66 minutes
-        let deadline = build_deadline(Some(too_long_ms));
-        assert!(deadline.is_some());
-        let deadline_time = deadline.unwrap();
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            // Should default to 10 minutes (600000 ms)
-            let expected_min = Duration::from_millis(DEFAULT_DURATION_MS) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(DEFAULT_DURATION_MS) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Duration above 1 hour should default to 10 minutes, got {duration:?}"
-            );
-        } else {
-            panic!("Default deadline should be in the future");
-        }
-
-        // Test 3: Valid duration (10 minutes) should pass through unchanged
-        let valid_ms = 10 * 60 * 1000u64; // 10 minutes
-        let deadline = build_deadline(Some(valid_ms));
-        assert!(deadline.is_some());
-        let deadline_time = deadline.unwrap();
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            let expected_min = Duration::from_millis(valid_ms) - Duration::from_secs(1);
-            let expected_max = Duration::from_millis(valid_ms) + Duration::from_secs(1);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Valid duration should pass through unchanged, got {duration:?}"
-            );
-        } else {
-            panic!("Valid deadline should be in the future");
-        }
-
-        // Test 4: Exactly at minimum (3 seconds) should pass through
-        let min_ms = 3_000u64; // Exactly 3 seconds
-        let deadline = build_deadline(Some(min_ms));
-        assert!(deadline.is_some());
-        let deadline_time = deadline.unwrap();
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            let expected_min = Duration::from_millis(min_ms) - Duration::from_millis(100);
-            let expected_max = Duration::from_millis(min_ms) + Duration::from_millis(100);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Duration at minimum should pass through, got {duration:?}"
-            );
-        } else {
-            panic!("Minimum deadline should be in the future");
-        }
-
-        // Test 5: Exactly at maximum (1 hour) should pass through
-        let max_ms = 3_600_000u64; // Exactly 1 hour
-        let deadline = build_deadline(Some(max_ms));
-        assert!(deadline.is_some());
-        let deadline_time = deadline.unwrap();
-        if let Ok(duration) = deadline_time.duration_since(now) {
-            let expected_min = Duration::from_millis(max_ms) - Duration::from_millis(1000);
-            let expected_max = Duration::from_millis(max_ms) + Duration::from_millis(1000);
-            assert!(
-                duration >= expected_min && duration <= expected_max,
-                "Duration at maximum should pass through, got {duration:?}"
-            );
-        } else {
-            panic!("Maximum deadline should be in the future");
-        }
-    }
-
-    /// Test that calculate_effective_timeout_ms applies the same logic as build_deadline.
-    /// This is critical for ensuring log_analysis_start displays the correct timeout.
-    #[test]
-    fn calculate_effective_timeout_ms_matches_build_deadline_logic() {
-        let now_ms = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .expect("SystemTime should be after UNIX_EPOCH")
-            .as_millis() as u64;
-
-        // Test 1: Relative duration (15 minutes) should pass through unchanged
-        let fifteen_minutes_ms = 15 * 60 * 1000u64;
-        let result = calculate_effective_timeout_ms(Some(fifteen_minutes_ms));
-        assert_eq!(
-            result,
-            Some(fifteen_minutes_ms),
-            "15 minute relative duration should pass through unchanged"
-        );
-
-        // Test 2: Absolute timestamp (now + 15 minutes) should convert to ~15 minutes
-        let absolute_15min = now_ms + fifteen_minutes_ms;
-        let result = calculate_effective_timeout_ms(Some(absolute_15min));
-        assert!(
-            result.is_some(),
-            "Future absolute timestamp should return Some"
-        );
-        let effective_ms = result.unwrap();
-        // Allow 2 second tolerance for timing variance
-        assert!(
-            effective_ms >= fifteen_minutes_ms - 2000 && effective_ms <= fifteen_minutes_ms + 2000,
-            "Absolute timestamp should convert to ~15 minutes, got {effective_ms}ms"
-        );
-
-        // Test 3: Duration below minimum (1 second) should default to 10 minutes
-        let too_short_ms = 1_000u64;
-        let result = calculate_effective_timeout_ms(Some(too_short_ms));
-        assert_eq!(
-            result,
-            Some(DEFAULT_DURATION_MS),
-            "Duration below 3 seconds should default to 10 minutes"
-        );
-
-        // Test 4: Duration above maximum (2 hours) should default to 10 minutes
-        let too_long_ms = 2 * 3_600_000u64;
-        let result = calculate_effective_timeout_ms(Some(too_long_ms));
-        assert_eq!(
-            result,
-            Some(DEFAULT_DURATION_MS),
-            "Duration above 1 hour should default to 10 minutes"
-        );
-
-        // Test 5: Past absolute timestamp should return None
-        let past_timestamp_ms = 1_700_000_000_000u64; // Circa late 2023
-        let result = calculate_effective_timeout_ms(Some(past_timestamp_ms));
-        assert!(
-            result.is_none(),
-            "Past timestamp should return None (deadline already passed)"
-        );
-
-        // Test 6: None should default to 10 minutes
-        let result = calculate_effective_timeout_ms(None);
-        assert_eq!(
-            result,
-            Some(DEFAULT_DURATION_MS),
-            "None should default to 10 minutes"
-        );
-    }
+    // Deadline tests moved to utils/deadline_tests.rs (Issue #268)
 
     #[test]
     fn record_cache_loads_once_per_neuron_under_contention() {
