+++ src/analysis/utils/deadline_tests.rs
@@ -1,0 +1,514 @@
+//! Tests for deadline and logging utilities (Issue #268).
+//!
+//! These tests were extracted from implementation.rs as part of the refactoring.
+
+use super::*;
+use std::time::{Duration, SystemTime};
+
+// ============================================================================
+// shuffle_within_top_k tests
+// ============================================================================
+
+#[test]
+fn shuffle_within_top_k_deterministic_with_seed() {
+    let mut a = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
+    let mut b = a.clone();
+    shuffle_within_top_k(&mut a[..], Some(123), "ctx", 5);
+    shuffle_within_top_k(&mut b[..], Some(123), "ctx", 5);
+    assert_eq!(a, b);
+}
+
+#[test]
+fn shuffle_within_top_k_only_shuffles_prefix() {
+    let mut items = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
+    shuffle_within_top_k(&mut items[..], Some(123), "ctx", 3);
+    // Suffix must remain untouched.
+    assert_eq!(items[3..], [4, 5, 6, 7, 8]);
+}
+
+#[test]
+fn shuffle_within_top_k_preserves_multiset() {
+    let mut items = [1u32, 2, 3, 4, 5, 6, 7, 8].to_vec();
+    let mut before = items.clone();
+    before.sort_unstable();
+    shuffle_within_top_k(&mut items[..], Some(999), "ctx", 6);
+    items.sort_unstable();
+    assert_eq!(items, before);
+}
+
+#[test]
+fn shuffle_within_top_k_no_op_when_top_k_is_zero() {
+    let mut items = [1u32, 2, 3, 4].to_vec();
+    let before = items.clone();
+    shuffle_within_top_k(&mut items[..], Some(1), "ctx", 0);
+    assert_eq!(items, before);
+}
+
+// ============================================================================
+// deadline_passed tests
+// ============================================================================
+
+#[test]
+fn deadline_passed_detects_elapsed_wall_clock_deadline() {
+    // Use the deadline override mechanism in tests so behaviour is deterministic
+    let _guard = deadline_override::DeadlineOverrideGuard::with_sequence(vec![true, false]);
+
+    let dummy_deadline = Some(SystemTime::now());
+    assert!(
+        deadline_passed(&dummy_deadline),
+        "past deadlines should be treated as expired immediately"
+    );
+
+    assert!(
+        !deadline_passed(&dummy_deadline),
+        "future deadlines should not be marked as expired"
+    );
+
+    assert!(
+        !deadline_passed(&None),
+        "missing deadlines should behave as if no timeout was requested"
+    );
+}
+
+// ============================================================================
+// build_deadline tests
+// ============================================================================
+
+#[test]
+fn build_deadline_handles_absolute_timestamps_and_relative_durations() {
+    // Verify that build_deadline correctly handles both absolute timestamps
+    // (milliseconds since UNIX_EPOCH) and relative durations (milliseconds from now)
+    let now = SystemTime::now();
+    let now_ms = now
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .expect("SystemTime should be after UNIX_EPOCH")
+        .as_millis() as u64;
+
+    // Test 1: Absolute timestamp (large value, >= year 2000)
+    // Create a deadline 10 minutes in the future using absolute timestamp
+    let ten_minutes_ms = 10 * 60 * 1000; // 10 minutes in milliseconds
+    let future_deadline_ms = now_ms + ten_minutes_ms;
+    let deadline = build_deadline(Some(future_deadline_ms));
+
+    assert!(
+        deadline.is_some(),
+        "deadline should be Some when deadline_ms is provided"
+    );
+
+    let deadline_time = deadline.unwrap();
+
+    // The deadline should be approximately 10 minutes in the future
+    // Allow for some small timing variance (up to 1 second)
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "absolute timestamp deadline should be approximately 10 minutes in the future, got {duration:?}"
+        );
+    } else {
+        panic!("deadline should be in the future");
+    }
+
+    // Test 2: Relative duration (small value, < year 2000)
+    // Pass 10 minutes as a relative duration
+    let relative_deadline_ms = ten_minutes_ms; // 10 minutes as relative duration
+    let relative_deadline = build_deadline(Some(relative_deadline_ms));
+    assert!(relative_deadline.is_some());
+    let relative_time = relative_deadline.unwrap();
+    // This should also be approximately 10 minutes in the future
+    if let Ok(duration) = relative_time.duration_since(now) {
+        let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "relative duration deadline should be approximately 10 minutes in the future, got {duration:?}"
+        );
+    } else {
+        panic!("relative deadline should be in the future");
+    }
+
+    // Test 3: Verify that an absolute timestamp in the past returns None
+    // (deadline already passed - no point in creating a deadline)
+    let past_timestamp_ms = 1_700_000_000_000u64; // Jan 2024 (in the past)
+    let past_deadline = build_deadline(Some(past_timestamp_ms));
+    assert!(
+        past_deadline.is_none(),
+        "Past timestamp should return None (deadline already passed)"
+    );
+
+    // Test 4: Verify that a future absolute timestamp is correctly converted to relative duration
+    let future_timestamp_ms = now_ms + ten_minutes_ms; // 10 minutes in the future as absolute timestamp
+    let future_deadline = build_deadline(Some(future_timestamp_ms));
+    assert!(future_deadline.is_some());
+    let future_time = future_deadline.unwrap();
+    // Should be approximately 10 minutes in the future
+    if let Ok(duration) = future_time.duration_since(now) {
+        let expected_min = Duration::from_millis(ten_minutes_ms) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(ten_minutes_ms) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Future absolute timestamp should be converted to relative duration correctly, got {duration:?}"
+        );
+    } else {
+        panic!("Future deadline should be in the future");
+    }
+}
+
+#[test]
+fn build_deadline_validates_duration_bounds() {
+    // Test that build_deadline validates and defaults to 10 minutes for invalid values
+    let now = SystemTime::now();
+
+    // Test 1: Duration below minimum (3 seconds) should default to 10 minutes
+    let too_short_ms = 1_000u64; // 1 second
+    let deadline = build_deadline(Some(too_short_ms));
+    assert!(deadline.is_some());
+    let deadline_time = deadline.unwrap();
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        // Should default to 10 minutes (600000 ms)
+        let expected_min = Duration::from_millis(DEFAULT_DURATION_MS) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(DEFAULT_DURATION_MS) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Duration below 3 seconds should default to 10 minutes, got {duration:?}"
+        );
+    } else {
+        panic!("Default deadline should be in the future");
+    }
+
+    // Test 2: Duration above maximum (1 hour) should default to 10 minutes
+    let too_long_ms = 4_000_000u64; // ~66 minutes
+    let deadline = build_deadline(Some(too_long_ms));
+    assert!(deadline.is_some());
+    let deadline_time = deadline.unwrap();
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        // Should default to 10 minutes (600000 ms)
+        let expected_min = Duration::from_millis(DEFAULT_DURATION_MS) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(DEFAULT_DURATION_MS) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Duration above 1 hour should default to 10 minutes, got {duration:?}"
+        );
+    } else {
+        panic!("Default deadline should be in the future");
+    }
+
+    // Test 3: Valid duration (10 minutes) should pass through unchanged
+    let valid_ms = 10 * 60 * 1000u64; // 10 minutes
+    let deadline = build_deadline(Some(valid_ms));
+    assert!(deadline.is_some());
+    let deadline_time = deadline.unwrap();
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        let expected_min = Duration::from_millis(valid_ms) - Duration::from_secs(1);
+        let expected_max = Duration::from_millis(valid_ms) + Duration::from_secs(1);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Valid duration should pass through unchanged, got {duration:?}"
+        );
+    } else {
+        panic!("Valid deadline should be in the future");
+    }
+
+    // Test 4: Exactly at minimum (3 seconds) should pass through
+    let min_ms = 3_000u64; // Exactly 3 seconds
+    let deadline = build_deadline(Some(min_ms));
+    assert!(deadline.is_some());
+    let deadline_time = deadline.unwrap();
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        let expected_min = Duration::from_millis(min_ms) - Duration::from_millis(100);
+        let expected_max = Duration::from_millis(min_ms) + Duration::from_millis(100);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Duration at minimum should pass through, got {duration:?}"
+        );
+    } else {
+        panic!("Minimum deadline should be in the future");
+    }
+
+    // Test 5: Exactly at maximum (1 hour) should pass through
+    let max_ms = 3_600_000u64; // Exactly 1 hour
+    let deadline = build_deadline(Some(max_ms));
+    assert!(deadline.is_some());
+    let deadline_time = deadline.unwrap();
+    if let Ok(duration) = deadline_time.duration_since(now) {
+        let expected_min = Duration::from_millis(max_ms) - Duration::from_millis(1000);
+        let expected_max = Duration::from_millis(max_ms) + Duration::from_millis(1000);
+        assert!(
+            duration >= expected_min && duration <= expected_max,
+            "Duration at maximum should pass through, got {duration:?}"
+        );
+    } else {
+        panic!("Maximum deadline should be in the future");
+    }
+}
+
+// ============================================================================
+// calculate_effective_timeout_ms tests
+// ============================================================================
+
+/// Test that calculate_effective_timeout_ms applies the same logic as build_deadline.
+/// This is critical for ensuring log_analysis_start displays the correct timeout.
+#[test]
+fn calculate_effective_timeout_ms_matches_build_deadline_logic() {
+    let now_ms = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .expect("SystemTime should be after UNIX_EPOCH")
+        .as_millis() as u64;
+
+    // Test 1: Relative duration (15 minutes) should pass through unchanged
+    let fifteen_minutes_ms = 15 * 60 * 1000u64;
+    let result = calculate_effective_timeout_ms(Some(fifteen_minutes_ms));
+    assert_eq!(
+        result,
+        Some(fifteen_minutes_ms),
+        "15 minute relative duration should pass through unchanged"
+    );
+
+    // Test 2: Absolute timestamp (now + 15 minutes) should convert to ~15 minutes
+    let absolute_15min = now_ms + fifteen_minutes_ms;
+    let result = calculate_effective_timeout_ms(Some(absolute_15min));
+    assert!(
+        result.is_some(),
+        "Future absolute timestamp should return Some"
+    );
+    let effective_ms = result.unwrap();
+    // Allow 2 second tolerance for timing variance
+    assert!(
+        effective_ms >= fifteen_minutes_ms - 2000 && effective_ms <= fifteen_minutes_ms + 2000,
+        "Absolute timestamp should convert to ~15 minutes, got {effective_ms}ms"
+    );
+
+    // Test 3: Duration below minimum (1 second) should default to 10 minutes
+    let too_short_ms = 1_000u64;
+    let result = calculate_effective_timeout_ms(Some(too_short_ms));
+    assert_eq!(
+        result,
+        Some(DEFAULT_DURATION_MS),
+        "Duration below 3 seconds should default to 10 minutes"
+    );
+
+    // Test 4: Duration above maximum (2 hours) should default to 10 minutes
+    let too_long_ms = 2 * 3_600_000u64;
+    let result = calculate_effective_timeout_ms(Some(too_long_ms));
+    assert_eq!(
+        result,
+        Some(DEFAULT_DURATION_MS),
+        "Duration above 1 hour should default to 10 minutes"
+    );
+
+    // Test 5: Past absolute timestamp should return None
+    let past_timestamp_ms = 1_700_000_000_000u64; // Circa late 2023
+    let result = calculate_effective_timeout_ms(Some(past_timestamp_ms));
+    assert!(
+        result.is_none(),
+        "Past timestamp should return None (deadline already passed)"
+    );
+
+    // Test 6: None should default to 10 minutes
+    let result = calculate_effective_timeout_ms(None);
+    assert_eq!(
+        result,
+        Some(DEFAULT_DURATION_MS),
+        "None should default to 10 minutes"
+    );
+}
+
+// ============================================================================
+// parse_input_index tests
+// ============================================================================
+
+#[test]
+fn parse_input_index_parses_valid_input_uuids() {
+    assert_eq!(parse_input_index("input-0"), Some(0));
+    assert_eq!(parse_input_index("input-42"), Some(42));
+    assert_eq!(parse_input_index("input-1000"), Some(1000));
+}
+
+#[test]
+fn parse_input_index_returns_none_for_invalid_uuids() {
+    assert_eq!(parse_input_index("hidden-123"), None);
+    assert_eq!(parse_input_index("output-0"), None);
+    assert_eq!(parse_input_index("input-"), None);
+    assert_eq!(parse_input_index("input-abc"), None);
+    assert_eq!(parse_input_index(""), None);
+}
+
+// ============================================================================
+// derive_seed tests
+// ============================================================================
+
+#[test]
+fn derive_seed_is_deterministic() {
+    let seed1 = derive_seed(12345, "context", 0);
+    let seed2 = derive_seed(12345, "context", 0);
+    assert_eq!(seed1, seed2);
+}
+
+#[test]
+fn derive_seed_varies_with_context() {
+    let seed1 = derive_seed(12345, "context1", 0);
+    let seed2 = derive_seed(12345, "context2", 0);
+    assert_ne!(seed1, seed2);
+}
+
+#[test]
+fn derive_seed_varies_with_salt() {
+    let seed1 = derive_seed(12345, "context", 0);
+    let seed2 = derive_seed(12345, "context", 1);
+    assert_ne!(seed1, seed2);
+}
+
+#[test]
+fn derive_seed_varies_with_base_seed() {
+    let seed1 = derive_seed(12345, "context", 0);
+    let seed2 = derive_seed(54321, "context", 0);
+    assert_ne!(seed1, seed2);
+}
+
+// ============================================================================
+// shuffle_slice tests
+// ============================================================================
+
+#[test]
+fn shuffle_slice_is_deterministic_with_seed() {
+    let mut a = [1, 2, 3, 4, 5, 6, 7, 8].to_vec();
+    let mut b = a.clone();
+    shuffle_slice(&mut a[..], Some(42), "test");
+    shuffle_slice(&mut b[..], Some(42), "test");
+    assert_eq!(a, b);
+}
+
+#[test]
+fn shuffle_slice_preserves_elements() {
+    let mut items = [1, 2, 3, 4, 5].to_vec();
+    let mut sorted = items.clone();
+    shuffle_slice(&mut items[..], Some(123), "test");
+    items.sort();
+    sorted.sort();
+    assert_eq!(items, sorted);
+}
+
+#[test]
+fn shuffle_slice_no_op_for_single_element() {
+    let mut items = [42].to_vec();
+    shuffle_slice(&mut items[..], Some(123), "test");
+    assert_eq!(items, [42]);
+}
+
+#[test]
+fn shuffle_slice_no_op_for_empty() {
+    let mut items: Vec<i32> = vec![];
+    shuffle_slice(&mut items[..], Some(123), "test");
+    assert!(items.is_empty());
+}
+
+// ============================================================================
+// calculate_gpu_batch_timeout tests
+// ============================================================================
+
+#[test]
+fn calculate_gpu_batch_timeout_returns_max_when_no_deadline() {
+    let timeout = calculate_gpu_batch_timeout(&None);
+    assert_eq!(timeout, Duration::from_secs(GPU_QUEUE_TIMEOUT_MAX_SECS));
+}
+
+#[test]
+fn calculate_gpu_batch_timeout_returns_min_when_deadline_passed() {
+    let past_deadline = Some(SystemTime::now() - Duration::from_secs(10));
+    let timeout = calculate_gpu_batch_timeout(&past_deadline);
+    assert_eq!(timeout, Duration::from_secs(GPU_QUEUE_TIMEOUT_MIN_SECS));
+}
+
+#[test]
+fn calculate_gpu_batch_timeout_uses_half_remaining_time() {
+    let future_deadline = Some(SystemTime::now() + Duration::from_secs(200));
+    let timeout = calculate_gpu_batch_timeout(&future_deadline);
+    // Half of 200 seconds is 100 seconds, which is between min (60) and max (300)
+    assert!(timeout >= Duration::from_secs(95) && timeout <= Duration::from_secs(105));
+}
+
+#[test]
+fn calculate_gpu_batch_timeout_clamps_to_min() {
+    let near_deadline = Some(SystemTime::now() + Duration::from_secs(30));
+    let timeout = calculate_gpu_batch_timeout(&near_deadline);
+    // Half of 30 seconds is 15, which is below min (60), so should return min
+    assert_eq!(timeout, Duration::from_secs(GPU_QUEUE_TIMEOUT_MIN_SECS));
+}
+
+#[test]
+fn calculate_gpu_batch_timeout_clamps_to_max() {
+    let far_deadline = Some(SystemTime::now() + Duration::from_secs(1000));
+    let timeout = calculate_gpu_batch_timeout(&far_deadline);
+    // Half of 1000 seconds is 500, which is above max (300), so should return max
+    assert_eq!(timeout, Duration::from_secs(GPU_QUEUE_TIMEOUT_MAX_SECS));
+}
+
+// ============================================================================
+// OrderedNeuron tests
+// ============================================================================
+
+#[test]
+fn ordered_neuron_stores_uuid_and_index() {
+    let neuron = OrderedNeuron {
+        uuid: "input-42".to_string(),
+        index: 42,
+    };
+    assert_eq!(neuron.uuid, "input-42");
+    assert_eq!(neuron.index, 42);
+}
+
+// ============================================================================
+// order_eligible_sources tests
+// ============================================================================
+
+#[test]
+fn order_eligible_sources_shuffles_deterministically_with_seed() {
+    let neurons: Vec<OrderedNeuron> = (0..10)
+        .map(|i| OrderedNeuron {
+            uuid: format!("hidden-{i}"),
+            index: i,
+        })
+        .collect();
+
+    let mut sources1: Vec<&OrderedNeuron> = neurons.iter().collect();
+    let mut sources2: Vec<&OrderedNeuron> = neurons.iter().collect();
+
+    order_eligible_sources(&mut sources1, Some(12345), "test", 0, None);
+    order_eligible_sources(&mut sources2, Some(12345), "test", 0, None);
+
+    let uuids1: Vec<&str> = sources1.iter().map(|n| n.uuid.as_str()).collect();
+    let uuids2: Vec<&str> = sources2.iter().map(|n| n.uuid.as_str()).collect();
+
+    assert_eq!(uuids1, uuids2);
+}
+
+#[test]
+fn order_eligible_sources_preserves_all_elements() {
+    let neurons: Vec<OrderedNeuron> = (0..5)
+        .map(|i| OrderedNeuron {
+            uuid: format!("neuron-{i}"),
+            index: i,
+        })
+        .collect();
+
+    let mut sources: Vec<&OrderedNeuron> = neurons.iter().collect();
+    order_eligible_sources(&mut sources, Some(42), "test", 0, None);
+
+    let mut uuids: Vec<&str> = sources.iter().map(|n| n.uuid.as_str()).collect();
+    uuids.sort();
+    let expected = ["neuron-0", "neuron-1", "neuron-2", "neuron-3", "neuron-4"];
+    assert_eq!(uuids, expected);
+}
+
+#[test]
+fn order_eligible_sources_no_op_for_single_element() {
+    let neurons = [OrderedNeuron {
+        uuid: "single".to_string(),
+        index: 0,
+    }];
+    let mut sources: Vec<&OrderedNeuron> = neurons.iter().collect();
+    order_eligible_sources(&mut sources, Some(42), "test", 0, None);
+    assert_eq!(sources.len(), 1);
+    assert_eq!(sources[0].uuid, "single");
+}
