+++ src/analysis/utils/mod.rs
@@ -7,7 +7,9 @@
 //!
 //! - `memory` - Memory detection and system requirements checking (Issue #267)
 //! - `platform` - Platform-specific setup (Linux XDG, Mesa warnings) (Issue #267)
+//! - `deadline` - Deadline handling and logging utilities (Issue #268)
 
+pub mod deadline;
 pub mod memory;
 pub mod platform;
 
@@ -21,6 +23,20 @@ pub use memory::{
 
 // Re-export platform setup functions
 pub use platform::{ensure_xdg_runtime_dir, suppress_mesa_warnings_if_requested};
+
+// Re-export deadline handling functions (Issue #268)
+pub use deadline::{
+    build_deadline, calculate_effective_timeout_ms, calculate_gpu_batch_timeout, deadline_passed,
+    derive_seed, focus_unused_observations_from_env, log_analysis_start, log_analysis_timeout,
+    order_eligible_sources, parse_input_index, shuffle_slice, shuffle_within_top_k,
+    source_input_index_bias_from_env, OrderedNeuron, DEFAULT_DURATION_MS,
+    GPU_QUEUE_TIMEOUT_MAX_SECS, GPU_QUEUE_TIMEOUT_MIN_SECS, MAX_DURATION_MS, MIN_DURATION_MS,
+    YEAR_2000_MS,
+};
+
+// Re-export deadline override for tests
+#[cfg(test)]
+pub use deadline::deadline_override;
 
 /// Check if verbose logging is enabled. Result is cached for performance.
 /// Set `NEAT_AI_DISCOVERY_VERBOSE=1` to enable verbose logging.
